refactor(meetings): split audio_process detector into focused modules
audio_process/mod.rs had grown to 2675 lines with a 342-line detection
loop and a 263-line state-machine match. Carve it into focused modules
and decompose the two giant functions, with no behavior change:

- model/state/resolve/suppress/lifecycle/events/tests modules, each with
  a single responsibility (mod.rs is now 257 lines)
- the loop drops to ~120 lines, reading as a sequence of named steps
  (handle_stop_signal, handle_auto_end_request, acquire_input_processes,
  build_candidates, advance_audio_process_state, apply_state_action)
- advance_audio_process_state's 5-arm match becomes a small dispatcher
  plus one advance_from_* function per state

Pure reorganization: no public surface change (only the run loop is
consumed externally) and all 154 meeting_watcher tests pass.
