+++ crates/screenpipe-engine/src/meeting_watcher/audio_process/mod.rs
@@ -13,7 +13,9 @@ use crate::meeting_watcher::shared::ignore::{
     meeting_app_is_ignored_with_terms, normalize_ignored_meeting_apps,
 };
 use crate::meeting_watcher::shared::profiles::{load_detection_profiles, MeetingDetectionProfile};
-use crate::meeting_watcher::shared::telemetry::{capture_detection_decision, capture_detection_outcome};
+use crate::meeting_watcher::shared::telemetry::{
+    capture_detection_decision, capture_detection_outcome,
+};
 use crate::routes::meetings::{emit_meeting_status_changed, resolve_meeting_status_from};
 use chrono::{DateTime, Utc};
 use futures::{FutureExt, StreamExt};
@@ -22,7 +24,7 @@ use screenpipe_audio::meeting_processes::{self, AudioInputProcess};
 use screenpipe_db::{DatabaseManager, MEETING_END_REASON_AUTO_END, MEETING_END_REASON_SHUTDOWN};
 use screenpipe_events::subscribe_to_event;
 use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
@@ -46,264 +48,23 @@ const ACTIVE_POLL_INTERVAL: Duration = Duration::from_secs(1);
 const IDLE_POLL_INTERVAL: Duration = Duration::from_secs(5);
 const UNKNOWN_BROWSER_PLATFORM: &str = "Unknown";
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub(crate) struct ProcessKey(String);
+mod model;
+pub(crate) use model::*;
 
-impl ProcessKey {
-    fn from_process(process: &AudioInputProcess) -> Option<Self> {
-        if let Some(session_id) = process
-            .audio_session_id
-            .as_deref()
-            .map(str::trim)
-            .filter(|id| !id.is_empty())
-        {
-            return Some(Self(format!("audio-session:{}", session_id.to_lowercase())));
-        }
-        if let Some(pid) = process.pid {
-            return Some(Self(format!("pid:{}", pid)));
-        }
-        process
-            .owner_bundle_id
-            .as_ref()
-            .or(process.bundle_id.as_ref())
-            .map(|id| Self(format!("bundle:{}", id.to_lowercase())))
-            .or_else(|| {
-                process
-                    .owner_app_name
-                    .as_ref()
-                    .or(process.process_name.as_ref())
-                    .map(|name| Self(format!("name:{}", name.to_lowercase())))
-            })
-    }
+mod events;
+pub(crate) use events::*;
 
-    fn reattached(platform: &str) -> Self {
-        Self(format!("reattached:{}", platform.to_lowercase()))
-    }
+mod suppress;
+pub(crate) use suppress::*;
 
-    fn is_reattached(&self) -> bool {
-        self.0.starts_with("reattached:")
-    }
-}
+mod lifecycle;
+pub(crate) use lifecycle::*;
 
-#[derive(Debug, Clone)]
-pub(crate) struct TrackedProcess {
-    pub first_seen_at: Instant,
-    pub last_seen_at: Instant,
-    pub raw: AudioInputProcess,
-}
+mod resolve;
+pub(crate) use resolve::*;
 
-#[derive(Debug)]
-pub(crate) struct StickyProcessTracker {
-    sticky_window: Duration,
-    tracked: HashMap<ProcessKey, TrackedProcess>,
-}
-
-impl StickyProcessTracker {
-    pub(crate) fn new(sticky_window: Duration) -> Self {
-        Self {
-            sticky_window,
-            tracked: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn update(
-        &mut self,
-        now: Instant,
-        processes: &[AudioInputProcess],
-    ) -> Vec<(ProcessKey, TrackedProcess)> {
-        let mut seen = HashSet::new();
-        for process in processes {
-            let Some(key) = ProcessKey::from_process(process) else {
-                continue;
-            };
-            seen.insert(key.clone());
-            self.tracked
-                .entry(key)
-                .and_modify(|tracked| {
-                    tracked.last_seen_at = now;
-                    tracked.raw = process.clone();
-                })
-                .or_insert_with(|| TrackedProcess {
-                    first_seen_at: now,
-                    last_seen_at: now,
-                    raw: process.clone(),
-                });
-        }
-
-        self.tracked.retain(|key, tracked| {
-            seen.contains(key) || now.duration_since(tracked.last_seen_at) <= self.sticky_window
-        });
-
-        let mut out: Vec<_> = self
-            .tracked
-            .iter()
-            .map(|(key, tracked)| (key.clone(), tracked.clone()))
-            .collect();
-        out.sort_by(|a, b| a.0.cmp(&b.0));
-        out
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct BrowserPageEvidence {
-    pub browser_app: Option<String>,
-    pub url: Option<String>,
-    pub title: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum ResolvedMeetingCandidate {
-    Native {
-        platform: String,
-        session_key: ProcessKey,
-        first_seen_at: Instant,
-        process: AudioInputProcess,
-    },
-    Browser {
-        platform: String,
-        meeting_url: String,
-        browser_app: String,
-        session_key: ProcessKey,
-        first_seen_at: Instant,
-        process: AudioInputProcess,
-    },
-    UnresolvedBrowser {
-        browser_app: String,
-        session_key: ProcessKey,
-        first_seen_at: Instant,
-        process: AudioInputProcess,
-    },
-    Ignored,
-    NonMeeting,
-}
-
-impl ResolvedMeetingCandidate {
-    fn resolved_session(&self) -> Option<ResolvedSession> {
-        match self {
-            Self::Native {
-                platform,
-                session_key,
-                first_seen_at,
-                ..
-            } => Some(ResolvedSession {
-                platform: platform.clone(),
-                session_key: session_key.clone(),
-                meeting_url: None,
-                first_seen_at: *first_seen_at,
-                is_browser: false,
-            }),
-            Self::Browser {
-                platform,
-                meeting_url,
-                session_key,
-                first_seen_at,
-                ..
-            } => Some(ResolvedSession {
-                platform: platform.clone(),
-                session_key: session_key.clone(),
-                meeting_url: Some(meeting_url.clone()),
-                first_seen_at: *first_seen_at,
-                is_browser: true,
-            }),
-            _ => None,
-        }
-    }
-
-    fn unresolved_browser_session(&self) -> Option<(String, ProcessKey, Instant)> {
-        match self {
-            Self::UnresolvedBrowser {
-                browser_app,
-                session_key,
-                first_seen_at,
-                ..
-            } => Some((browser_app.clone(), session_key.clone(), *first_seen_at)),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct AxResolvedCandidate {
-    pub browser_app: String,
-    pub profile_index: usize,
-    pub meeting_url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ResolvedSession {
-    platform: String,
-    session_key: ProcessKey,
-    meeting_url: Option<String>,
-    first_seen_at: Instant,
-    is_browser: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct SuppressedSession {
-    session_key: ProcessKey,
-    platform: Option<String>,
-    meeting_url: Option<String>,
-}
-
-#[allow(dead_code)]
-#[derive(Debug)]
-pub(crate) enum AudioProcessMeetingState {
-    Idle,
-    Candidate {
-        platform: String,
-        session_key: ProcessKey,
-        meeting_url: Option<String>,
-        first_seen_at: Instant,
-        is_browser: bool,
-    },
-    CandidateUnresolvedBrowser {
-        browser_app: String,
-        session_key: ProcessKey,
-        first_seen_at: Instant,
-        last_resolution_attempt: Instant,
-    },
-    Active {
-        meeting_id: i64,
-        platform: String,
-        session_key: ProcessKey,
-        meeting_url: Option<String>,
-        first_seen_at: Instant,
-        last_seen_at: Instant,
-        is_browser: bool,
-    },
-    Ending {
-        meeting_id: i64,
-        platform: String,
-        session_key: ProcessKey,
-        meeting_url: Option<String>,
-        first_seen_at: Instant,
-        since: Instant,
-        is_browser: bool,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum AudioProcessStateAction {
-    StartMeeting {
-        platform: String,
-        session_key: ProcessKey,
-        meeting_url: Option<String>,
-        first_seen_at: Instant,
-        is_browser: bool,
-    },
-    EndMeeting {
-        meeting_id: i64,
-        suppressed_session: Option<SuppressedSession>,
-    },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AutoStartOutcome {
-    Started(i64),
-    AdoptedActive(i64),
-    BlockedByActive(i64),
-    Failed,
-}
+mod state;
+pub(crate) use state::*;
 
 /// Run the audio-process-first meeting detection loop.
 pub async fn run_audio_process_meeting_detection_loop(
@@ -374,18 +135,7 @@ pub async fn run_audio_process_meeting_detection_loop(
             _ = tokio::time::sleep(interval) => {}
             _ = shutdown_rx.recv() => {
                 info!("audio-process meeting detector: shutdown received");
-                if let Some(meeting_id) = active_or_ending_meeting_id(&state) {
-                    let now = Utc::now().format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string();
-                    if let Err(e) = db
-                        .end_meeting(meeting_id, &now, Some(MEETING_END_REASON_SHUTDOWN))
-                        .await
-                    {
-                        error!(
-                            "audio-process meeting detector: failed to end meeting on shutdown: {}",
-                            e
-                        );
-                    }
-                }
+                end_active_meeting_on_shutdown(&db, &state).await;
                 sync_meeting_flag(false, &in_meeting_flag, &detector);
                 return;
             }
@@ -396,67 +146,27 @@ pub async fn run_audio_process_meeting_detection_loop(
         }
 
         if let Some(event) = stop_sub.next().now_or_never().flatten() {
-            let stop_signal = event.data;
-            if active_or_ending_meeting_id(&state) == Some(stop_signal.meeting_id) {
-                if let Some(session) =
-                    active_or_ending_suppressed_session(&state, stop_signal.meeting_id)
-                {
-                    suppress_session(&mut suppressed_sessions, session);
-                }
-                info!(
-                    "audio-process meeting detector: explicit stop moved detector to idle (meeting_id={})",
-                    stop_signal.meeting_id
-                );
-                state = AudioProcessMeetingState::Idle;
-                last_explicit_stop_id = Some(stop_signal.meeting_id);
-                sync_meeting_flag(false, &in_meeting_flag, &detector);
-            }
+            handle_stop_signal(
+                event.data,
+                &mut state,
+                &mut suppressed_sessions,
+                &mut last_explicit_stop_id,
+                &in_meeting_flag,
+                &detector,
+            );
         }
 
         if let Some(event) = auto_end_sub.next().now_or_never().flatten() {
-            let request = event.data;
-            let manual_matches = { *manual_meeting.read().await == Some(request.meeting_id) };
-            let detector_matches = active_or_ending_meeting_id(&state) == Some(request.meeting_id);
-            if manual_matches || detector_matches {
-                let now = Utc::now().format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string();
-                match db
-                    .end_meeting_with_typed_text(
-                        request.meeting_id,
-                        &now,
-                        false,
-                        Some(MEETING_END_REASON_AUTO_END),
-                    )
-                    .await
-                {
-                    Ok(()) => {
-                        debug!(
-                            "audio-process meeting detector: auto-ended inactive live meeting (id={}, reason={})",
-                            request.meeting_id,
-                            request.reason.as_deref().unwrap_or("unknown")
-                        );
-                        if manual_matches {
-                            let mut manual = manual_meeting.write().await;
-                            if *manual == Some(request.meeting_id) {
-                                *manual = None;
-                            }
-                        }
-                        if detector_matches {
-                            if let Some(session) =
-                                active_or_ending_suppressed_session(&state, request.meeting_id)
-                            {
-                                suppress_session(&mut suppressed_sessions, session);
-                            }
-                            state = AudioProcessMeetingState::Idle;
-                        }
-                        sync_meeting_flag(false, &in_meeting_flag, &detector);
-                        emit_ended_and_status(&db, &manual_meeting, request.meeting_id).await;
-                    }
-                    Err(e) => warn!(
-                        "audio-process meeting detector: failed to auto-end inactive live meeting {}: {}",
-                        request.meeting_id, e
-                    ),
-                }
-            }
+            handle_auto_end_request(
+                event.data,
+                &db,
+                &manual_meeting,
+                &mut state,
+                &mut suppressed_sessions,
+                &in_meeting_flag,
+                &detector,
+            )
+            .await;
         }
 
         {
@@ -475,64 +185,31 @@ pub async fn run_audio_process_meeting_detection_loop(
             continue;
         }
 
-        let snapshot = meeting_processes::current_input_processes();
-        if !snapshot.supported {
-            if !unsupported_logged {
-                info!(
-                    "audio-process meeting detector: unsupported platform ({})",
-                    snapshot.error.as_deref().unwrap_or("no details")
-                );
-                unsupported_logged = true;
+        let processes = match acquire_input_processes(
+            &mut unsupported_logged,
+            &state,
+            &in_meeting_flag,
+            &detector,
+        ) {
+            Some(processes) => processes,
+            None => {
+                interval = IDLE_POLL_INTERVAL;
+                continue;
             }
-            sync_meeting_flag(
-                matches!(state, AudioProcessMeetingState::Active { .. }),
-                &in_meeting_flag,
-                &detector,
-            );
-            interval = IDLE_POLL_INTERVAL;
-            continue;
-        }
-        if let Some(error) = snapshot.error.as_deref() {
-            debug!("audio-process meeting detector: snapshot error: {}", error);
-        }
-
-        let processes: Vec<_> = snapshot
-            .processes
-            .into_iter()
-            .filter(|process| process.pid != Some(std::process::id() as i32))
-            .collect();
-        let now = Instant::now();
-        let live_session_keys = live_process_keys(&processes);
-        retain_live_suppressed_sessions(&mut suppressed_sessions, &live_session_keys);
-        let tracked = tracker.update(now, &processes);
-        let live_tracked: Vec<_> = tracked
-            .iter()
-            .filter(|(key, _)| live_session_keys.contains(key))
-            .cloned()
-            .collect();
-        let ax_candidates = if needs_ax_resolution(&state) {
-            should_use_ax_fallback(&live_tracked, &profiles, &ignored_terms).await
-        } else {
-            Vec::new()
         };
-        let mut candidates = resolve_tracked_candidates(
+        let now = Instant::now();
+
+        let (candidates, live_candidates) = build_candidates(
             &db,
             &profiles,
             &ignored_terms,
-            &tracked,
-            ax_candidates,
+            &processes,
+            &mut tracker,
+            &mut suppressed_sessions,
+            &state,
+            now,
         )
         .await;
-
-        candidates.retain(|candidate| !matches!(candidate, ResolvedMeetingCandidate::Ignored));
-        filter_suppressed_candidates(&mut candidates, &suppressed_sessions);
-        let live_candidates: Vec<_> = candidates
-            .iter()
-            .filter(|candidate| {
-                candidate_session_key(candidate).is_some_and(|key| live_session_keys.contains(key))
-            })
-            .cloned()
-            .collect();
 
         let was_active = matches!(state, AudioProcessMeetingState::Active { .. });
         let was_ending = matches!(state, AudioProcessMeetingState::Ending { .. });
@@ -550,93 +227,20 @@ pub async fn run_audio_process_meeting_detection_loop(
         state = new_state;
 
         if let Some(action) = action {
-            match action {
-                AudioProcessStateAction::StartMeeting {
-                    platform,
-                    session_key,
-                    meeting_url,
-                    first_seen_at,
-                    is_browser,
-                } => {
-                    let (cal_title, cal_attendees) =
-                        find_overlapping_calendar_event(&calendar_events);
-                    let attendees_str = cal_attendees.as_ref().map(|a| a.join(", "));
-                    let outcome = start_or_adopt_auto_meeting(
-                        &db,
-                        &manual_meeting,
-                        &platform,
-                        cal_title.as_deref(),
-                        attendees_str.as_deref(),
-                        last_explicit_stop_id,
-                    )
-                    .await;
-                    match outcome {
-                        AutoStartOutcome::Started(meeting_id)
-                        | AutoStartOutcome::AdoptedActive(meeting_id) => {
-                            state = AudioProcessMeetingState::Active {
-                                meeting_id,
-                                platform: platform.clone(),
-                                session_key,
-                                meeting_url,
-                                first_seen_at,
-                                last_seen_at: now,
-                                is_browser,
-                            };
-                            sync_meeting_flag(true, &in_meeting_flag, &detector);
-                            if let Ok(status) =
-                                resolve_meeting_status_from(db.as_ref(), manual_meeting.as_ref())
-                                    .await
-                            {
-                                emit_meeting_status_changed(&status);
-                            }
-                        }
-                        AutoStartOutcome::BlockedByActive(meeting_id) => {
-                            debug!(
-                                "audio-process meeting detector: start blocked by already-active meeting {}; waiting",
-                                meeting_id
-                            );
-                            state = AudioProcessMeetingState::Idle;
-                            sync_meeting_flag(true, &in_meeting_flag, &detector);
-                            if let Ok(status) =
-                                resolve_meeting_status_from(db.as_ref(), manual_meeting.as_ref())
-                                    .await
-                            {
-                                emit_meeting_status_changed(&status);
-                            }
-                        }
-                        AutoStartOutcome::Failed => {
-                            state = AudioProcessMeetingState::Idle;
-                        }
-                    }
-                    flap_count = 0;
-                }
-                AudioProcessStateAction::EndMeeting {
-                    meeting_id,
-                    suppressed_session,
-                } => {
-                    if let Some(session) = suppressed_session {
-                        suppress_session(&mut suppressed_sessions, session);
-                    }
-                    let now_ts = Utc::now().format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string();
-                    match db
-                        .end_meeting_with_typed_text(meeting_id, &now_ts, true, None)
-                        .await
-                    {
-                        Ok(()) => {
-                            capture_meeting_outcome(db.as_ref(), meeting_id, flap_count).await;
-                            flap_count = 0;
-                            emit_ended_and_status(&db, &manual_meeting, meeting_id).await;
-                        }
-                        Err(e) => {
-                            error!(
-                                "audio-process meeting detector: failed to end meeting {}: {}",
-                                meeting_id, e
-                            );
-                        }
-                    }
-                    sync_meeting_flag(false, &in_meeting_flag, &detector);
-                }
-            }
+            apply_state_action(
+                action,
+                &db,
+                &manual_meeting,
+                &mut state,
+                &mut suppressed_sessions,
+                &mut flap_count,
+                &in_meeting_flag,
+                &detector,
+                last_explicit_stop_id,
+                &calendar_events,
+                now,
+            )
+            .await;
         }
 
         let active_now = matches!(state, AudioProcessMeetingState::Active { .. });
@@ -649,2027 +253,5 @@ pub async fn run_audio_process_meeting_detection_loop(
     }
 }
 
-async fn resolve_tracked_candidates(
-    db: &DatabaseManager,
-    profiles: &[MeetingDetectionProfile],
-    ignored_terms: &[String],
-    tracked: &[(ProcessKey, TrackedProcess)],
-    ax_candidates: Vec<AxResolvedCandidate>,
-) -> Vec<ResolvedMeetingCandidate> {
-    let evidence = match db_find_browser_evidence(db).await {
-        Ok(evidence) => evidence,
-        Err(e) => {
-            debug!(
-                "audio-process meeting detector: DB browser evidence query failed: {}",
-                e
-            );
-            Vec::new()
-        }
-    };
-
-    tracked
-        .iter()
-        .map(|(key, tracked)| {
-            resolve_process_candidate(
-                key.clone(),
-                tracked.first_seen_at,
-                &tracked.raw,
-                profiles,
-                &evidence,
-                &ax_candidates,
-                ignored_terms,
-            )
-        })
-        .collect()
-}
-
-/// AX resolution (a full accessibility-tree walk of every browser window) is
-/// only useful while we're still trying to attribute a browser to a known
-/// meeting platform. Once a meeting is `Active` — or already winding down in
-/// `Ending` — the platform is settled, so re-walking the tree every
-/// `ACTIVE_POLL_INTERVAL` for the rest of the call is pure overhead.
-fn needs_ax_resolution(state: &AudioProcessMeetingState) -> bool {
-    !matches!(
-        state,
-        AudioProcessMeetingState::Active { .. } | AudioProcessMeetingState::Ending { .. }
-    )
-}
-
-async fn should_use_ax_fallback(
-    tracked: &[(ProcessKey, TrackedProcess)],
-    profiles: &[MeetingDetectionProfile],
-    ignored_terms: &[String],
-) -> Vec<AxResolvedCandidate> {
-    let has_browser = tracked
-        .iter()
-        .any(|(_, tracked)| browser_app_name(&tracked.raw).is_some());
-    if !has_browser {
-        return Vec::new();
-    }
-
-    ax_resolved_candidates(profiles, ignored_terms).await
-}
-
-
-async fn db_find_browser_evidence(
-    db: &DatabaseManager,
-) -> Result<Vec<BrowserPageEvidence>, sqlx::Error> {
-    // `frames.timestamp` is RFC3339 (`...T...+00:00`). Comparing it to
-    // SQLite's `datetime()` string (`... ...`) is lexical and pulls in stale
-    // same-day frames.
-    let rows: Vec<(String, String, Option<String>)> = sqlx::query_as(
-        "SELECT DISTINCT app_name, window_name, browser_url FROM frames \
-         WHERE timestamp > strftime('%Y-%m-%dT%H:%M:%f+00:00', 'now', '-10 seconds') \
-         AND app_name IS NOT NULL AND window_name IS NOT NULL",
-    )
-    .fetch_all(&db.pool)
-    .await?;
-
-    Ok(rows
-        .into_iter()
-        .filter(|(app_name, _, _)| is_browser_app(app_name))
-        .map(|(app_name, window_name, browser_url)| BrowserPageEvidence {
-            browser_app: Some(app_name),
-            url: browser_url,
-            title: Some(window_name),
-        })
-        .collect())
-}
-
-pub(crate) fn resolve_process_candidate(
-    session_key: ProcessKey,
-    first_seen_at: Instant,
-    process: &AudioInputProcess,
-    profiles: &[MeetingDetectionProfile],
-    evidence: &[BrowserPageEvidence],
-    ax_candidates: &[AxResolvedCandidate],
-    ignored_terms: &[String],
-) -> ResolvedMeetingCandidate {
-    if let Some(browser_app) = browser_app_name(process) {
-        if let Some((profile_index, meeting_url)) =
-            resolve_browser_evidence(&browser_app, profiles, evidence)
-        {
-            let profile = &profiles[profile_index];
-            let platform = platform_name_for_profile(profile, true);
-            if candidate_is_ignored(
-                &platform,
-                Some(profile),
-                process,
-                ignored_terms,
-                Some(&browser_app),
-                Some(&meeting_url),
-                None,
-            ) {
-                return ResolvedMeetingCandidate::Ignored;
-            }
-            return ResolvedMeetingCandidate::Browser {
-                platform,
-                meeting_url,
-                browser_app,
-                session_key,
-                first_seen_at,
-                process: process.clone(),
-            };
-        }
-
-        if let Some(ax) = resolve_ax_browser_candidate(&browser_app, profiles, ax_candidates) {
-            let profile = &profiles[ax.profile_index];
-            let platform = platform_name_for_profile(profile, true);
-            let meeting_url = ax.meeting_url.unwrap_or_else(|| platform.clone());
-            if candidate_is_ignored(
-                &platform,
-                Some(profile),
-                process,
-                ignored_terms,
-                Some(&browser_app),
-                Some(&meeting_url),
-                None,
-            ) {
-                return ResolvedMeetingCandidate::Ignored;
-            }
-            return ResolvedMeetingCandidate::Browser {
-                platform,
-                meeting_url,
-                browser_app,
-                session_key,
-                first_seen_at,
-                process: process.clone(),
-            };
-        }
-
-        if candidate_is_ignored(
-            UNKNOWN_BROWSER_PLATFORM,
-            None,
-            process,
-            ignored_terms,
-            Some(&browser_app),
-            None,
-            None,
-        ) {
-            return ResolvedMeetingCandidate::Ignored;
-        }
-
-        return ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app,
-            session_key,
-            first_seen_at,
-            process: process.clone(),
-        };
-    }
-
-    if let Some((platform, profile_index)) = resolve_native_platform(process, profiles) {
-        let profile = profile_index.and_then(|idx| profiles.get(idx));
-        if candidate_is_ignored(&platform, profile, process, ignored_terms, None, None, None) {
-            return ResolvedMeetingCandidate::Ignored;
-        }
-        return ResolvedMeetingCandidate::Native {
-            platform,
-            session_key,
-            first_seen_at,
-            process: process.clone(),
-        };
-    }
-
-    ResolvedMeetingCandidate::NonMeeting
-}
-
-fn resolve_browser_evidence(
-    browser_app: &str,
-    profiles: &[MeetingDetectionProfile],
-    evidence: &[BrowserPageEvidence],
-) -> Option<(usize, String)> {
-    for item in evidence {
-        if let Some(app) = item.browser_app.as_deref() {
-            if !browser_names_match(browser_app, app) {
-                continue;
-            }
-        }
-        for (idx, profile) in profiles.iter().enumerate() {
-            if profile.app_identifiers.browser_url_patterns.is_empty()
-                && profile.app_identifiers.browser_title_patterns.is_empty()
-            {
-                continue;
-            }
-            if browser_window_matches_meeting(item.url.as_deref(), item.title.as_deref(), profile) {
-                let meeting_url = item
-                    .url
-                    .clone()
-                    .or_else(|| item.title.clone())
-                    .unwrap_or_else(|| platform_name_for_profile(profile, true));
-                return Some((idx, meeting_url));
-            }
-        }
-    }
-    None
-}
-
-fn resolve_ax_browser_candidate(
-    browser_app: &str,
-    profiles: &[MeetingDetectionProfile],
-    ax_candidates: &[AxResolvedCandidate],
-) -> Option<AxResolvedCandidate> {
-    ax_candidates.iter().find_map(|candidate| {
-        if !browser_names_match(browser_app, &candidate.browser_app) {
-            return None;
-        }
-        profiles.get(candidate.profile_index)?;
-        Some(candidate.clone())
-    })
-}
-
-fn resolve_native_platform(
-    process: &AudioInputProcess,
-    profiles: &[MeetingDetectionProfile],
-) -> Option<(String, Option<usize>)> {
-    let fields = process_identity_fields(process);
-    for field in &fields {
-        if let Some(platform) = known_native_bundle_platform(field) {
-            return Some((platform.to_string(), None));
-        }
-    }
-
-    for (idx, profile) in profiles.iter().enumerate() {
-        let matches = profile.app_identifiers.macos_app_names.iter().any(|name| {
-            fields
-                .iter()
-                .any(|field| field.eq_ignore_ascii_case(name) || field == &name.to_lowercase())
-        });
-        if matches {
-            return Some((platform_name_for_profile(profile, false), Some(idx)));
-        }
-    }
-    None
-}
-
-fn process_identity_fields(process: &AudioInputProcess) -> Vec<String> {
-    [
-        process.bundle_id.as_deref(),
-        process.owner_bundle_id.as_deref(),
-        process.process_name.as_deref(),
-        process.owner_app_name.as_deref(),
-    ]
-    .into_iter()
-    .flatten()
-    .map(|s| s.trim().to_lowercase())
-    .filter(|s| !s.is_empty())
-    .collect()
-}
-
-fn known_native_bundle_platform(field_lower: &str) -> Option<&'static str> {
-    // Matching is over identity fields that are macOS bundle ids on macOS and
-    // Windows exe names on Windows (e.g. `Zoom.exe`, `ms-teams.exe`), so the arms
-    // below accept both forms.
-    if field_lower.starts_with("us.zoom.")
-        || field_lower == "zoom.us"
-        || field_lower == "zoom"
-        || field_lower == "zoom.exe"
-    {
-        return Some("Zoom");
-    }
-    if field_lower.starts_with("com.microsoft.teams")
-        || field_lower == "microsoft teams"
-        || field_lower == "teams"
-        // Windows: ms-teams.exe, teams.exe, ms-teams_modulehost.exe
-        || field_lower.starts_with("ms-teams")
-        || field_lower == "teams.exe"
-    {
-        return Some("Microsoft Teams");
-    }
-    if field_lower.starts_with("com.tinyspeck.slackmacgap")
-        || field_lower == "slack"
-        || field_lower == "slack.exe"
-    {
-        return Some("Slack");
-    }
-    if field_lower == "com.apple.facetime" || field_lower == "facetime" {
-        return Some("FaceTime");
-    }
-    if field_lower.contains("webex") {
-        return Some("Webex");
-    }
-    if field_lower.contains("discord") {
-        return Some("Discord");
-    }
-    if field_lower.contains("signal") {
-        return Some("Signal");
-    }
-    if field_lower.contains("whatsapp") {
-        return Some("WhatsApp");
-    }
-    if field_lower.contains("telegram") {
-        return Some("Telegram");
-    }
-    if field_lower.contains("skype") {
-        return Some("Skype");
-    }
-    None
-}
-
-fn browser_app_name(process: &AudioInputProcess) -> Option<String> {
-    let bundle = process
-        .owner_bundle_id
-        .as_deref()
-        .or(process.bundle_id.as_deref());
-    if let Some(bundle) = bundle {
-        if let Some(name) = browser_name_for_bundle(bundle) {
-            return Some(name.to_string());
-        }
-    }
-
-    [
-        process.owner_app_name.as_deref(),
-        process.process_name.as_deref(),
-        process.bundle_id.as_deref(),
-    ]
-    .into_iter()
-    .flatten()
-    .find(|name| is_browser_app(name))
-    .map(normalize_browser_display_name)
-}
-
-fn browser_name_for_bundle(bundle: &str) -> Option<&'static str> {
-    let bundle = bundle.to_lowercase();
-    if bundle.starts_with("com.google.chrome") || bundle.starts_with("com.google.chrome.helper") {
-        Some("Google Chrome")
-    } else if bundle == "com.apple.safari" || bundle.starts_with("com.apple.safari.") {
-        Some("Safari")
-    } else if bundle.starts_with("company.thebrowser.browser") {
-        Some("Arc")
-    } else if bundle.starts_with("org.mozilla.firefox") {
-        Some("Firefox")
-    } else if bundle.starts_with("com.microsoft.edgemac") {
-        Some("Microsoft Edge")
-    } else if bundle.starts_with("com.brave.browser") {
-        Some("Brave Browser")
-    } else if bundle.starts_with("org.chromium.chromium") {
-        Some("Chromium")
-    } else if bundle.contains("opera") {
-        Some("Opera")
-    } else if bundle.contains("vivaldi") {
-        Some("Vivaldi")
-    } else if bundle.contains("zen") {
-        Some("Zen Browser")
-    } else {
-        None
-    }
-}
-
-fn normalize_browser_display_name(name: &str) -> String {
-    let lower = name.to_lowercase();
-    if lower.contains("chrome") {
-        "Google Chrome".to_string()
-    } else if lower.contains("safari") {
-        "Safari".to_string()
-    } else if lower.contains("arc") {
-        "Arc".to_string()
-    } else if lower.contains("firefox") {
-        "Firefox".to_string()
-    } else if lower.contains("edge") || lower.contains("msedge") {
-        "Microsoft Edge".to_string()
-    } else if lower.contains("brave") {
-        "Brave Browser".to_string()
-    } else {
-        name.to_string()
-    }
-}
-
-fn browser_names_match(left: &str, right: &str) -> bool {
-    let left = normalize_browser_display_name(left).to_lowercase();
-    let right = normalize_browser_display_name(right).to_lowercase();
-    left == right || left.contains(&right) || right.contains(&left)
-}
-
-fn platform_name_for_profile(profile: &MeetingDetectionProfile, browser: bool) -> String {
-    let ids = &profile.app_identifiers;
-    let has_pattern = |needle: &str| {
-        ids.browser_url_patterns
-            .iter()
-            .any(|pattern| pattern.contains(needle))
-            || ids
-                .browser_title_patterns
-                .iter()
-                .any(|pattern| pattern.contains(needle))
-    };
-
-    if has_pattern("meet.google.com") {
-        "Google Meet".to_string()
-    } else if has_pattern("zoom.us") {
-        "Zoom".to_string()
-    } else if has_pattern("teams.microsoft.com") || has_pattern("teams.live.com") {
-        "Microsoft Teams".to_string()
-    } else if has_pattern("app.slack.com/huddle") {
-        if browser {
-            "Slack Huddle".to_string()
-        } else {
-            "Slack".to_string()
-        }
-    } else if has_pattern("discord.com") {
-        "Discord".to_string()
-    } else if has_pattern("daily.co") || has_pattern("app.daily.co") {
-        "Daily".to_string()
-    } else if has_pattern("webex.com") {
-        "Webex".to_string()
-    } else if has_pattern("web.whatsapp.com") {
-        "WhatsApp".to_string()
-    } else if has_pattern("web.telegram.org") {
-        "Telegram".to_string()
-    } else if ids.macos_app_names.iter().any(|n| *n == "facetime") {
-        "FaceTime".to_string()
-    } else if ids.macos_app_names.iter().any(|n| *n == "signal") {
-        "Signal".to_string()
-    } else if let Some(name) = ids.macos_app_names.first() {
-        title_case_ascii(name)
-    } else if let Some(pattern) = ids.browser_url_patterns.first() {
-        title_case_ascii(pattern.split('/').next().unwrap_or(pattern))
-    } else {
-        "Meeting".to_string()
-    }
-}
-
-fn title_case_ascii(value: &str) -> String {
-    value
-        .split([' ', '.', '-'])
-        .filter(|part| !part.is_empty())
-        .map(|part| {
-            let mut chars = part.chars();
-            match chars.next() {
-                Some(first) => format!("{}{}", first.to_ascii_uppercase(), chars.as_str()),
-                None => String::new(),
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn candidate_is_ignored(
-    platform: &str,
-    profile: Option<&MeetingDetectionProfile>,
-    process: &AudioInputProcess,
-    ignored_terms: &[String],
-    browser_app: Option<&str>,
-    meeting_url: Option<&str>,
-    title: Option<&str>,
-) -> bool {
-    if ignored_terms.is_empty() {
-        return false;
-    }
-    if let Some(profile) = profile {
-        if meeting_app_is_ignored_with_terms(platform, profile, ignored_terms) {
-            return true;
-        }
-    }
-
-    ignored_terms.iter().any(|term| {
-        contains_normalized_term(platform, term)
-            || browser_app.is_some_and(|app| contains_normalized_term(app, term))
-            || meeting_url.is_some_and(|url| contains_normalized_term(url, term))
-            || title.is_some_and(|title| contains_normalized_term(title, term))
-            || process_identity_fields(process)
-                .iter()
-                .any(|field| contains_normalized_term(field, term))
-    })
-}
-
-pub(crate) fn advance_audio_process_state(
-    state: AudioProcessMeetingState,
-    live_candidates: &[ResolvedMeetingCandidate],
-    session_candidates: &[ResolvedMeetingCandidate],
-    now: Instant,
-    confirm_window: Duration,
-    ending_grace: Duration,
-) -> (AudioProcessMeetingState, Option<AudioProcessStateAction>) {
-    let resolved = live_candidates
-        .iter()
-        .find_map(ResolvedMeetingCandidate::resolved_session);
-    let unresolved = live_candidates
-        .iter()
-        .find_map(ResolvedMeetingCandidate::unresolved_browser_session);
-
-    match state {
-        AudioProcessMeetingState::Idle => {
-            if let Some(session) = resolved {
-                (
-                    AudioProcessMeetingState::Candidate {
-                        platform: session.platform,
-                        session_key: session.session_key,
-                        meeting_url: session.meeting_url,
-                        first_seen_at: session.first_seen_at,
-                        is_browser: session.is_browser,
-                    },
-                    None,
-                )
-            } else if let Some((browser_app, session_key, first_seen_at)) = unresolved {
-                (
-                    AudioProcessMeetingState::CandidateUnresolvedBrowser {
-                        browser_app,
-                        session_key,
-                        first_seen_at,
-                        last_resolution_attempt: now,
-                    },
-                    None,
-                )
-            } else {
-                (AudioProcessMeetingState::Idle, None)
-            }
-        }
-        AudioProcessMeetingState::Candidate {
-            platform,
-            session_key,
-            meeting_url,
-            first_seen_at,
-            is_browser,
-        } => {
-            if let Some(session) = resolved {
-                if session.session_key == session_key
-                    && session.platform == platform
-                    && session.meeting_url == meeting_url
-                {
-                    if now.duration_since(first_seen_at) >= confirm_window {
-                        return (
-                            AudioProcessMeetingState::Active {
-                                meeting_id: -1,
-                                platform: platform.clone(),
-                                session_key: session_key.clone(),
-                                meeting_url: meeting_url.clone(),
-                                first_seen_at,
-                                last_seen_at: now,
-                                is_browser,
-                            },
-                            Some(AudioProcessStateAction::StartMeeting {
-                                platform,
-                                session_key,
-                                meeting_url,
-                                first_seen_at,
-                                is_browser,
-                            }),
-                        );
-                    }
-                    (
-                        AudioProcessMeetingState::Candidate {
-                            platform,
-                            session_key,
-                            meeting_url,
-                            first_seen_at,
-                            is_browser,
-                        },
-                        None,
-                    )
-                } else {
-                    (
-                        AudioProcessMeetingState::Candidate {
-                            platform: session.platform,
-                            session_key: session.session_key,
-                            meeting_url: session.meeting_url,
-                            first_seen_at: session.first_seen_at,
-                            is_browser: session.is_browser,
-                        },
-                        None,
-                    )
-                }
-            } else if let Some((browser_app, key, _)) = unresolved {
-                if key == session_key {
-                    (
-                        AudioProcessMeetingState::CandidateUnresolvedBrowser {
-                            browser_app,
-                            session_key,
-                            first_seen_at,
-                            last_resolution_attempt: now,
-                        },
-                        None,
-                    )
-                } else {
-                    (AudioProcessMeetingState::Idle, None)
-                }
-            } else {
-                (AudioProcessMeetingState::Idle, None)
-            }
-        }
-        AudioProcessMeetingState::CandidateUnresolvedBrowser {
-            browser_app,
-            session_key,
-            first_seen_at,
-            ..
-        } => {
-            if let Some(session) = resolved {
-                (
-                    AudioProcessMeetingState::Candidate {
-                        platform: session.platform,
-                        session_key: session.session_key,
-                        meeting_url: session.meeting_url,
-                        first_seen_at: now,
-                        is_browser: session.is_browser,
-                    },
-                    None,
-                )
-            } else if let Some((next_browser_app, key, next_first_seen_at)) = unresolved {
-                if key == session_key {
-                    // A browser holding the mic is NOT sufficient evidence to
-                    // auto-start a recorded meeting: voice notes, dictation, and
-                    // arbitrary WebRTC sites all hold the mic. We only start once
-                    // the browser resolves to a known platform/URL (the `resolved`
-                    // branch above), so an unresolved browser stays a pending
-                    // candidate indefinitely while we keep attempting resolution.
-                    let _ = confirm_window;
-                    (
-                        AudioProcessMeetingState::CandidateUnresolvedBrowser {
-                            browser_app: next_browser_app,
-                            session_key,
-                            first_seen_at,
-                            last_resolution_attempt: now,
-                        },
-                        None,
-                    )
-                } else {
-                    (
-                        AudioProcessMeetingState::CandidateUnresolvedBrowser {
-                            browser_app: next_browser_app,
-                            session_key: key,
-                            first_seen_at: next_first_seen_at,
-                            last_resolution_attempt: now,
-                        },
-                        None,
-                    )
-                }
-            } else {
-                let _ = browser_app;
-                (AudioProcessMeetingState::Idle, None)
-            }
-        }
-        AudioProcessMeetingState::Active {
-            meeting_id,
-            platform,
-            session_key,
-            meeting_url,
-            first_seen_at,
-            is_browser,
-            ..
-        } => {
-            if session_present(
-                session_candidates,
-                &session_key,
-                &platform,
-                meeting_url.as_deref(),
-            ) {
-                (
-                    AudioProcessMeetingState::Active {
-                        meeting_id,
-                        platform,
-                        session_key,
-                        meeting_url,
-                        first_seen_at,
-                        last_seen_at: now,
-                        is_browser,
-                    },
-                    None,
-                )
-            } else {
-                (
-                    AudioProcessMeetingState::Ending {
-                        meeting_id,
-                        platform,
-                        session_key,
-                        meeting_url,
-                        first_seen_at,
-                        since: now,
-                        is_browser,
-                    },
-                    None,
-                )
-            }
-        }
-        AudioProcessMeetingState::Ending {
-            meeting_id,
-            platform,
-            session_key,
-            meeting_url,
-            first_seen_at,
-            since,
-            is_browser,
-        } => {
-            if session_present(
-                session_candidates,
-                &session_key,
-                &platform,
-                meeting_url.as_deref(),
-            ) {
-                (
-                    AudioProcessMeetingState::Active {
-                        meeting_id,
-                        platform,
-                        session_key,
-                        meeting_url,
-                        first_seen_at,
-                        last_seen_at: now,
-                        is_browser,
-                    },
-                    None,
-                )
-            } else if now.duration_since(since) >= ending_grace {
-                (
-                    AudioProcessMeetingState::Idle,
-                    Some(AudioProcessStateAction::EndMeeting {
-                        meeting_id,
-                        suppressed_session: Some(SuppressedSession {
-                            session_key,
-                            platform: Some(platform),
-                            meeting_url,
-                        }),
-                    }),
-                )
-            } else {
-                (
-                    AudioProcessMeetingState::Ending {
-                        meeting_id,
-                        platform,
-                        session_key,
-                        meeting_url,
-                        first_seen_at,
-                        since,
-                        is_browser,
-                    },
-                    None,
-                )
-            }
-        }
-    }
-}
-
-fn session_present(
-    candidates: &[ResolvedMeetingCandidate],
-    session_key: &ProcessKey,
-    platform: &str,
-    meeting_url: Option<&str>,
-) -> bool {
-    // A reattached meeting (adopted from the DB after a restart) has a synthetic
-    // session key and no real process/url to reconcile against — and a browser
-    // meeting (e.g. Google Meet) can take several poll cycles to re-resolve its
-    // platform after restart, surfacing as an `UnresolvedBrowser` in the gap.
-    // Keep it alive while *any* live meeting session is present; genuine
-    // disappearance still flows through the normal ending grace.
-    if session_key.is_reattached() {
-        return candidates.iter().any(|candidate| {
-            matches!(
-                candidate,
-                ResolvedMeetingCandidate::Native { .. }
-                    | ResolvedMeetingCandidate::Browser { .. }
-                    | ResolvedMeetingCandidate::UnresolvedBrowser { .. }
-            )
-        });
-    }
-    // Non-reattached sessions must match on the real process key (the
-    // `is_reattached()` branch above is the only place a synthetic key matches).
-    candidates.iter().any(|candidate| match candidate {
-        ResolvedMeetingCandidate::Native {
-            platform: candidate_platform,
-            session_key: key,
-            ..
-        } => meeting_url.is_none() && key == session_key && candidate_platform == platform,
-        // A confirmed browser meeting is keyed by its audio session. Once it is
-        // live, the same browser audio session still holding the mic is sufficient
-        // proof the call is ongoing — so keep it alive on the `session_key` alone.
-        // We intentionally do NOT require the platform/URL to keep re-resolving:
-        // browser URL/platform attribution comes from a rolling 10s window of
-        // screen-capture evidence, which lapses whenever the user switches tabs or
-        // apps, the call goes fullscreen, or capture simply misses the URL. When
-        // that happens the candidate degrades to `UnresolvedBrowser` (or surfaces a
-        // slightly different URL string), but the audio session is unchanged. The
-        // mic is only released when the user actually leaves the call, which drops
-        // the `session_key` and lets the normal ending grace fire. URL/platform
-        // strictness still gates *starting* a meeting; this check only keeps a
-        // already-live one alive.
-        ResolvedMeetingCandidate::Browser {
-            session_key: key, ..
-        }
-        | ResolvedMeetingCandidate::UnresolvedBrowser {
-            session_key: key, ..
-        } => key == session_key,
-        _ => false,
-    })
-}
-
-fn candidate_session_key(candidate: &ResolvedMeetingCandidate) -> Option<&ProcessKey> {
-    match candidate {
-        ResolvedMeetingCandidate::Native { session_key, .. }
-        | ResolvedMeetingCandidate::Browser { session_key, .. }
-        | ResolvedMeetingCandidate::UnresolvedBrowser { session_key, .. } => Some(session_key),
-        _ => None,
-    }
-}
-
-fn filter_suppressed_candidates(
-    candidates: &mut Vec<ResolvedMeetingCandidate>,
-    suppressed_sessions: &[SuppressedSession],
-) {
-    if suppressed_sessions.is_empty() {
-        return;
-    }
-    candidates.retain(|candidate| !candidate_is_suppressed(candidate, suppressed_sessions));
-}
-
-fn candidate_is_suppressed(
-    candidate: &ResolvedMeetingCandidate,
-    suppressed_sessions: &[SuppressedSession],
-) -> bool {
-    suppressed_sessions
-        .iter()
-        .any(|session| session.matches_candidate(candidate))
-}
-
-fn retain_live_suppressed_sessions(
-    suppressed_sessions: &mut Vec<SuppressedSession>,
-    live_session_keys: &HashSet<ProcessKey>,
-) {
-    if suppressed_sessions.is_empty() {
-        return;
-    }
-    suppressed_sessions.retain(|session| live_session_keys.contains(&session.session_key));
-}
-
-fn live_process_keys(processes: &[AudioInputProcess]) -> HashSet<ProcessKey> {
-    processes
-        .iter()
-        .filter_map(ProcessKey::from_process)
-        .collect()
-}
-
-fn suppress_session(suppressed_sessions: &mut Vec<SuppressedSession>, session: SuppressedSession) {
-    if !suppressed_sessions.contains(&session) {
-        suppressed_sessions.push(session);
-    }
-}
-
-impl SuppressedSession {
-    fn matches_candidate(&self, candidate: &ResolvedMeetingCandidate) -> bool {
-        match candidate {
-            ResolvedMeetingCandidate::Native {
-                platform,
-                session_key,
-                ..
-            } => {
-                self.meeting_url.is_none()
-                    && self.session_key == *session_key
-                    && self.platform.as_deref().is_none_or(|p| p == platform)
-            }
-            ResolvedMeetingCandidate::Browser {
-                platform,
-                meeting_url,
-                session_key,
-                ..
-            } => {
-                self.session_key == *session_key
-                    && self.platform.as_deref().is_none_or(|p| p == platform)
-                    && self
-                        .meeting_url
-                        .as_deref()
-                        .is_none_or(|url| url == meeting_url)
-            }
-            ResolvedMeetingCandidate::UnresolvedBrowser { session_key, .. } => {
-                self.meeting_url.is_none() && self.session_key == *session_key
-            }
-            _ => false,
-        }
-    }
-}
-
-async fn start_or_adopt_auto_meeting(
-    db: &DatabaseManager,
-    manual_meeting: &tokio::sync::RwLock<Option<i64>>,
-    platform: &str,
-    title: Option<&str>,
-    attendees: Option<&str>,
-    last_explicit_stop_id: Option<i64>,
-) -> AutoStartOutcome {
-    if let Some(id) = *manual_meeting.read().await {
-        debug!(
-            "audio-process meeting detector: not starting {}, manual meeting {} is active",
-            platform, id
-        );
-        return AutoStartOutcome::BlockedByActive(id);
-    }
-
-    match db.get_most_recent_active_meeting().await {
-        Ok(Some(active)) => {
-            debug!(
-                "audio-process meeting detector: not starting {}, active meeting {} already exists",
-                platform, active.id
-            );
-            return AutoStartOutcome::BlockedByActive(active.id);
-        }
-        Ok(None) => {}
-        Err(e) => {
-            warn!(
-                "audio-process meeting detector: failed active meeting check before start: {}",
-                e
-            );
-            return AutoStartOutcome::Failed;
-        }
-    }
-
-    let merge_candidate = match db.find_recent_meeting_for_app(platform, 120).await {
-        Ok(Some(recent)) if last_explicit_stop_id == Some(recent.id) => Ok(None),
-        other => other,
-    };
-
-    match merge_candidate {
-        Ok(Some(recent)) => match db.reopen_meeting(recent.id).await {
-            Ok(()) => {
-                if title.is_some() && recent.title.as_ref().is_none_or(|t| t.is_empty()) {
-                    if let Err(e) = db
-                        .update_meeting(recent.id, None, None, title, attendees, None, None)
-                        .await
-                    {
-                        warn!(
-                            "audio-process meeting detector: failed to enrich reopened meeting {}: {}",
-                            recent.id, e
-                        );
-                    }
-                }
-                if let Ok(meeting) = db.get_meeting_by_id(recent.id).await {
-                    capture_detection_decision(&meeting, "audio_process_reopen", None);
-                }
-                AutoStartOutcome::AdoptedActive(recent.id)
-            }
-            Err(e) => {
-                warn!(
-                    "audio-process meeting detector: failed to reopen meeting {}: {}",
-                    recent.id, e
-                );
-                insert_new_audio_process_meeting(db, platform, title, attendees).await
-            }
-        },
-        Ok(None) => insert_new_audio_process_meeting(db, platform, title, attendees).await,
-        Err(e) => {
-            warn!(
-                "audio-process meeting detector: failed to find recent meeting: {}",
-                e
-            );
-            insert_new_audio_process_meeting(db, platform, title, attendees).await
-        }
-    }
-}
-
-async fn insert_new_audio_process_meeting(
-    db: &DatabaseManager,
-    platform: &str,
-    title: Option<&str>,
-    attendees: Option<&str>,
-) -> AutoStartOutcome {
-    match db
-        .insert_meeting(platform, "audio_process", title, attendees)
-        .await
-    {
-        Ok(id) => {
-            info!(
-                "audio-process meeting detector: meeting started (id={}, app={})",
-                id, platform
-            );
-            if let Err(e) = screenpipe_events::send_event(
-                "meeting_started",
-                serde_json::json!({
-                    "meeting_id": id,
-                    "app": platform,
-                    "title": title,
-                    "detection_source": "audio_process",
-                }),
-            ) {
-                warn!(
-                    "audio-process meeting detector: failed to emit meeting_started event: {}",
-                    e
-                );
-            }
-            if let Ok(meeting) = db.get_meeting_by_id(id).await {
-                capture_detection_decision(&meeting, "audio_process_start", None);
-            }
-            AutoStartOutcome::Started(id)
-        }
-        Err(e) => {
-            let msg = e.to_string();
-            if msg.contains("UNIQUE constraint failed") && msg.contains("idx_meetings_single_open")
-            {
-                match db.get_most_recent_active_meeting().await {
-                    Ok(Some(active)) => AutoStartOutcome::BlockedByActive(active.id),
-                    _ => AutoStartOutcome::Failed,
-                }
-            } else {
-                error!(
-                    "audio-process meeting detector: failed to insert meeting: {}",
-                    e
-                );
-                AutoStartOutcome::Failed
-            }
-        }
-    }
-}
-
-async fn emit_ended_and_status(
-    db: &DatabaseManager,
-    manual_meeting: &tokio::sync::RwLock<Option<i64>>,
-    meeting_id: i64,
-) {
-    if let Err(e) = screenpipe_events::send_event(
-        "meeting_ended",
-        serde_json::json!({ "meeting_id": meeting_id }),
-    ) {
-        warn!(
-            "audio-process meeting detector: failed to emit meeting_ended event: {}",
-            e
-        );
-    }
-    if let Ok(status) = resolve_meeting_status_from(db, manual_meeting).await {
-        emit_meeting_status_changed(&status);
-    }
-}
-
-async fn capture_meeting_outcome(db: &DatabaseManager, meeting_id: i64, flap_count: u32) {
-    if let Ok(meeting) = db.get_meeting_by_id(meeting_id).await {
-        capture_detection_outcome(&meeting, "auto_timeout", flap_count);
-    }
-}
-
-fn active_or_ending_meeting_id(state: &AudioProcessMeetingState) -> Option<i64> {
-    match state {
-        AudioProcessMeetingState::Active { meeting_id, .. }
-        | AudioProcessMeetingState::Ending { meeting_id, .. }
-            if *meeting_id >= 0 =>
-        {
-            Some(*meeting_id)
-        }
-        _ => None,
-    }
-}
-
-fn active_or_ending_suppressed_session(
-    state: &AudioProcessMeetingState,
-    expected_meeting_id: i64,
-) -> Option<SuppressedSession> {
-    match state {
-        AudioProcessMeetingState::Active {
-            meeting_id,
-            platform,
-            session_key,
-            meeting_url,
-            ..
-        }
-        | AudioProcessMeetingState::Ending {
-            meeting_id,
-            platform,
-            session_key,
-            meeting_url,
-            ..
-        } if *meeting_id == expected_meeting_id => Some(SuppressedSession {
-            session_key: session_key.clone(),
-            platform: Some(platform.clone()),
-            meeting_url: meeting_url.clone(),
-        }),
-        _ => None,
-    }
-}
-
-fn is_active_ending_flap(
-    was_active: bool,
-    was_ending: bool,
-    next: &AudioProcessMeetingState,
-) -> bool {
-    let now_active = matches!(next, AudioProcessMeetingState::Active { .. });
-    let now_ending = matches!(next, AudioProcessMeetingState::Ending { .. });
-    (was_active && now_ending) || (was_ending && now_active)
-}
-
-fn sync_meeting_flag(in_meeting: bool, flag: &AtomicBool, detector: &Option<Arc<MeetingDetector>>) {
-    flag.store(in_meeting, Ordering::Relaxed);
-    if let Some(detector) = detector {
-        detector.set_v2_in_meeting(in_meeting);
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct CalendarEventSignal {
-    pub title: String,
-    pub start: String,
-    pub end: String,
-    #[serde(default)]
-    pub attendees: Vec<String>,
-    #[serde(default)]
-    pub is_all_day: bool,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct MeetingAutoEndRequest {
-    meeting_id: i64,
-    #[serde(default)]
-    reason: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct DetectorStopSignal {
-    pub meeting_id: i64,
-    pub app: String,
-}
-
-fn find_overlapping_calendar_event(
-    events: &[CalendarEventSignal],
-) -> (Option<String>, Option<Vec<String>>) {
-    let now = Utc::now();
-    for cal_event in events {
-        if let (Ok(start), Ok(end)) = (
-            DateTime::parse_from_rfc3339(&cal_event.start),
-            DateTime::parse_from_rfc3339(&cal_event.end),
-        ) {
-            let start_utc = start.with_timezone(&Utc);
-            let end_utc = end.with_timezone(&Utc);
-            if start_utc <= now && end_utc >= now {
-                return (
-                    Some(cal_event.title.clone()),
-                    if cal_event.attendees.is_empty() {
-                        None
-                    } else {
-                        Some(cal_event.attendees.clone())
-                    },
-                );
-            }
-        }
-    }
-    (None, None)
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use screenpipe_db::DatabaseManager;
-
-    fn chrome_process() -> AudioInputProcess {
-        AudioInputProcess {
-            audio_session_id: Some("coreaudio-process:100:input:built-in-mic".to_string()),
-            audio_object_id: Some(100),
-            pid: Some(42),
-            bundle_id: Some("com.google.Chrome.helper".to_string()),
-            process_name: Some("Google Chrome Helper".to_string()),
-            owner_app_name: Some("Google Chrome".to_string()),
-            owner_bundle_id: Some("com.google.Chrome".to_string()),
-            first_seen_at_ms: None,
-        }
-    }
-
-    fn arc_process() -> AudioInputProcess {
-        AudioInputProcess {
-            audio_session_id: Some("coreaudio-process:300:input:built-in-mic".to_string()),
-            audio_object_id: Some(300),
-            pid: Some(84),
-            bundle_id: Some("company.thebrowser.Browser.helper".to_string()),
-            process_name: Some("Arc Helper".to_string()),
-            owner_app_name: Some("Arc".to_string()),
-            owner_bundle_id: Some("company.thebrowser.Browser".to_string()),
-            first_seen_at_ms: None,
-        }
-    }
-
-    fn zoom_process() -> AudioInputProcess {
-        AudioInputProcess {
-            audio_session_id: Some("coreaudio-process:200:input:built-in-mic".to_string()),
-            audio_object_id: Some(200),
-            pid: Some(7),
-            bundle_id: Some("us.zoom.xos".to_string()),
-            process_name: Some("zoom.us".to_string()),
-            owner_app_name: Some("zoom.us".to_string()),
-            owner_bundle_id: Some("us.zoom.xos".to_string()),
-            first_seen_at_ms: None,
-        }
-    }
-
-    fn migrated_db_url(name: &str) -> (tempfile::TempDir, String) {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join(name);
-        (dir, path.to_string_lossy().to_string())
-    }
-
-    async fn setup_db() -> (tempfile::TempDir, DatabaseManager) {
-        let (dir, path) = migrated_db_url("audio-process-meetings.db");
-        let db = DatabaseManager::new(&path, Default::default())
-            .await
-            .unwrap();
-        sqlx::migrate!("../screenpipe-db/src/migrations")
-            .run(&db.pool)
-            .await
-            .unwrap();
-        (dir, db)
-    }
-
-    #[test]
-    fn sticky_tracker_preserves_first_seen_across_short_drop() {
-        let start = Instant::now();
-        let mut tracker = StickyProcessTracker::new(Duration::from_secs(4));
-        let first = tracker.update(start, &[zoom_process()]);
-        let first_seen = first[0].1.first_seen_at;
-
-        let dropped = tracker.update(start + Duration::from_secs(2), &[]);
-        assert_eq!(dropped.len(), 1);
-        assert_eq!(dropped[0].1.first_seen_at, first_seen);
-
-        let returned = tracker.update(start + Duration::from_secs(3), &[zoom_process()]);
-        assert_eq!(returned.len(), 1);
-        assert_eq!(returned[0].1.first_seen_at, first_seen);
-    }
-
-    #[test]
-    fn sticky_tracker_evicts_after_window() {
-        let start = Instant::now();
-        let mut tracker = StickyProcessTracker::new(Duration::from_secs(4));
-        let first = tracker.update(start, &[zoom_process()]);
-        let first_seen = first[0].1.first_seen_at;
-
-        assert!(tracker
-            .update(start + Duration::from_secs(5), &[])
-            .is_empty());
-
-        let returned = tracker.update(start + Duration::from_secs(6), &[zoom_process()]);
-        assert_ne!(returned[0].1.first_seen_at, first_seen);
-    }
-
-    #[test]
-    fn native_bundle_maps_to_platform() {
-        let profiles = load_detection_profiles();
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&zoom_process()).unwrap(),
-            Instant::now(),
-            &zoom_process(),
-            &profiles,
-            &[],
-            &[],
-            &[],
-        );
-        assert!(matches!(
-            candidate,
-            ResolvedMeetingCandidate::Native { platform, .. } if platform == "Zoom"
-        ));
-    }
-
-    #[test]
-    fn windows_exe_names_map_to_native_platform() {
-        // Windows snapshots carry exe basenames (from the WASAPI sensor) rather
-        // than macOS bundle ids; the matcher must resolve both.
-        for (exe, expected) in [
-            ("zoom.exe", "Zoom"),
-            ("ms-teams.exe", "Microsoft Teams"),
-            ("teams.exe", "Microsoft Teams"),
-            ("ms-teams_modulehost.exe", "Microsoft Teams"),
-            ("slack.exe", "Slack"),
-            ("webex.exe", "Webex"),
-        ] {
-            assert_eq!(
-                known_native_bundle_platform(exe),
-                Some(expected),
-                "{exe} should resolve to {expected}"
-            );
-        }
-    }
-
-    #[test]
-    fn browser_helper_alone_is_unresolved_browser() {
-        let profiles = load_detection_profiles();
-        let process = chrome_process();
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&process).unwrap(),
-            Instant::now(),
-            &process,
-            &profiles,
-            &[],
-            &[],
-            &[],
-        );
-        assert!(matches!(
-            candidate,
-            ResolvedMeetingCandidate::UnresolvedBrowser { browser_app, .. }
-                if browser_app == "Google Chrome"
-        ));
-    }
-
-    #[test]
-    fn browser_helper_plus_meet_url_resolves_to_google_meet() {
-        let profiles = load_detection_profiles();
-        let process = chrome_process();
-        let evidence = vec![BrowserPageEvidence {
-            browser_app: Some("Google Chrome".to_string()),
-            url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-            title: Some("Meet".to_string()),
-        }];
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&process).unwrap(),
-            Instant::now(),
-            &process,
-            &profiles,
-            &evidence,
-            &[],
-            &[],
-        );
-        assert!(matches!(
-            candidate,
-            ResolvedMeetingCandidate::Browser { platform, meeting_url, .. }
-                if platform == "Google Meet" && meeting_url.contains("meet.google.com")
-        ));
-    }
-
-    #[test]
-    fn browser_helper_unrelated_tab_does_not_auto_start() {
-        let profiles = load_detection_profiles();
-        let process = chrome_process();
-        let evidence = vec![BrowserPageEvidence {
-            browser_app: Some("Google Chrome".to_string()),
-            url: Some("https://calendar.google.com/calendar/u/0/r".to_string()),
-            title: Some("Join with Google Meet - Calendar".to_string()),
-        }];
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&process).unwrap(),
-            Instant::now(),
-            &process,
-            &profiles,
-            &evidence,
-            &[],
-            &[],
-        );
-        assert!(matches!(
-            candidate,
-            ResolvedMeetingCandidate::UnresolvedBrowser { .. }
-        ));
-        let start = Instant::now();
-        let candidates = vec![candidate];
-        let (state, action) = advance_audio_process_state(
-            AudioProcessMeetingState::Idle,
-            &candidates,
-            &candidates,
-            start,
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        let (_state, second_action) = advance_audio_process_state(
-            state,
-            &candidates,
-            &candidates,
-            start + Duration::from_secs(10),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        // An unattributed browser tab holding the mic (here a Calendar tab) must
-        // never auto-start a recorded "Unknown" meeting: it stays a pending
-        // candidate, waiting to resolve to a known platform/URL.
-        assert!(action.is_none());
-        assert!(second_action.is_none());
-        assert!(matches!(
-            _state,
-            AudioProcessMeetingState::CandidateUnresolvedBrowser { .. }
-        ));
-    }
-
-    #[test]
-    fn arc_slack_client_url_does_not_auto_start_when_unattributed() {
-        let profiles = load_detection_profiles();
-        let process = arc_process();
-        let evidence = vec![BrowserPageEvidence {
-            browser_app: Some("Arc".to_string()),
-            url: Some("https://app.slack.com/client/T0BBNEEH6Q2/D0BBT87MELU".to_string()),
-            title: Some(
-                "Alex N (DM) - 1651 Market Apartments Residents - 1 new item - Slack".to_string(),
-            ),
-        }];
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&process).unwrap(),
-            Instant::now(),
-            &process,
-            &profiles,
-            &evidence,
-            &[],
-            &[],
-        );
-        assert!(
-            matches!(candidate, ResolvedMeetingCandidate::UnresolvedBrowser { .. }),
-            "Slack web client URLs are not yet attribution evidence, but Arc mic activity must remain startable"
-        );
-
-        let start = Instant::now();
-        let candidates = vec![candidate];
-        let (state, action) = advance_audio_process_state(
-            AudioProcessMeetingState::Idle,
-            &candidates,
-            &candidates,
-            start,
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(action.is_none());
-
-        let (_state, action) = advance_audio_process_state(
-            state,
-            &candidates,
-            &candidates,
-            start + Duration::from_secs(3),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        // A Slack web-client tab is not meeting evidence; the unattributed
-        // browser must not auto-start a recorded meeting.
-        assert!(action.is_none());
-        assert!(matches!(
-            _state,
-            AudioProcessMeetingState::CandidateUnresolvedBrowser { .. }
-        ));
-    }
-
-    #[test]
-    fn ax_fallback_can_resolve_browser_platform() {
-        let profiles = load_detection_profiles();
-        let process = chrome_process();
-        let meet_profile = profiles
-            .iter()
-            .position(|profile| {
-                profile
-                    .app_identifiers
-                    .browser_url_patterns
-                    .contains(&"meet.google.com")
-            })
-            .unwrap();
-        let ax = vec![AxResolvedCandidate {
-            browser_app: "Google Chrome".to_string(),
-            profile_index: meet_profile,
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-        }];
-        let candidate = resolve_process_candidate(
-            ProcessKey::from_process(&process).unwrap(),
-            Instant::now(),
-            &process,
-            &profiles,
-            &[],
-            &ax,
-            &[],
-        );
-        assert!(matches!(
-            candidate,
-            ResolvedMeetingCandidate::Browser { platform, .. } if platform == "Google Meet"
-        ));
-    }
-
-    #[test]
-    fn unresolved_browser_does_not_start_after_confirmation() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let candidate = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-        let (state, _) = advance_audio_process_state(
-            AudioProcessMeetingState::Idle,
-            std::slice::from_ref(&candidate),
-            std::slice::from_ref(&candidate),
-            start,
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        let (_state, action) = advance_audio_process_state(
-            state,
-            std::slice::from_ref(&candidate),
-            std::slice::from_ref(&candidate),
-            start + Duration::from_secs(3),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        // Holding the mic in an unresolved browser past the confirm window must
-        // not auto-start a meeting; we only start once it resolves to a known
-        // platform/URL.
-        assert!(action.is_none());
-        assert!(matches!(
-            _state,
-            AudioProcessMeetingState::CandidateUnresolvedBrowser { .. }
-        ));
-    }
-
-    #[test]
-    fn unresolved_browser_keeps_unknown_meeting_alive() {
-        let process = arc_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let active = AudioProcessMeetingState::Active {
-            meeting_id: 123,
-            platform: UNKNOWN_BROWSER_PLATFORM.to_string(),
-            session_key: key.clone(),
-            meeting_url: None,
-            first_seen_at: start,
-            last_seen_at: start,
-            is_browser: true,
-        };
-        let unresolved = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Arc".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-
-        let (state, action) = advance_audio_process_state(
-            active,
-            std::slice::from_ref(&unresolved),
-            std::slice::from_ref(&unresolved),
-            start + Duration::from_secs(1),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-
-        assert!(matches!(
-            state,
-            AudioProcessMeetingState::Active {
-                platform,
-                meeting_url: None,
-                ..
-            } if platform == UNKNOWN_BROWSER_PLATFORM
-        ));
-        assert!(action.is_none());
-    }
-
-    #[test]
-    fn resolved_browser_meeting_stays_alive_when_url_evidence_lost() {
-        // Regression: a confirmed Google Meet (browser) must NOT auto-end just
-        // because screen-capture URL evidence lapsed — e.g. the user switched to
-        // another tab/app, the call went fullscreen, or capture simply missed the
-        // URL within the 10s evidence window. The same browser audio session is
-        // still holding the mic, which is sufficient proof the call is ongoing.
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let active = AudioProcessMeetingState::Active {
-            meeting_id: 7,
-            platform: "Google Meet".to_string(),
-            session_key: key.clone(),
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-            first_seen_at: start,
-            last_seen_at: start,
-            is_browser: true,
-        };
-        // URL evidence lapsed this poll: the same browser audio session still
-        // holds the mic, but it no longer resolves to a known platform/URL.
-        let unresolved = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-
-        let (state, action) = advance_audio_process_state(
-            active,
-            std::slice::from_ref(&unresolved),
-            std::slice::from_ref(&unresolved),
-            start + Duration::from_secs(1),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-
-        assert!(
-            matches!(state, AudioProcessMeetingState::Active { .. }),
-            "resolved browser meeting must stay Active while the same audio session holds the mic, got {state:?}"
-        );
-        assert!(action.is_none());
-    }
-
-    #[test]
-    fn sticky_process_absent_from_live_snapshot_cannot_start() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let sticky_only = vec![ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/abc-defg-hij".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        }];
-
-        let (state, action) = advance_audio_process_state(
-            AudioProcessMeetingState::Idle,
-            &[],
-            &sticky_only,
-            start + Duration::from_secs(10),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-
-        assert!(matches!(state, AudioProcessMeetingState::Idle));
-        assert!(action.is_none());
-    }
-
-    #[test]
-    fn unresolved_browser_resolution_gets_fresh_confirm_window() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let unresolved = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Google Chrome".to_string(),
-            session_key: key.clone(),
-            first_seen_at: start,
-            process: process.clone(),
-        };
-        let (state, _) = advance_audio_process_state(
-            AudioProcessMeetingState::Idle,
-            std::slice::from_ref(&unresolved),
-            std::slice::from_ref(&unresolved),
-            start,
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-
-        let resolved = ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/abc-defg-hij".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-        let resolved_at = start + Duration::from_secs(10);
-        let (state, action) = advance_audio_process_state(
-            state,
-            std::slice::from_ref(&resolved),
-            std::slice::from_ref(&resolved),
-            resolved_at,
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(action.is_none());
-        assert!(matches!(
-            &state,
-            AudioProcessMeetingState::Candidate { first_seen_at, .. } if *first_seen_at == resolved_at
-        ));
-
-        let (_state, action) = advance_audio_process_state(
-            state,
-            std::slice::from_ref(&resolved),
-            std::slice::from_ref(&resolved),
-            resolved_at + Duration::from_secs(3),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(
-            action,
-            Some(AudioProcessStateAction::StartMeeting { first_seen_at, .. }) if first_seen_at == resolved_at
-        ));
-    }
-
-    #[test]
-    fn browser_meeting_ends_only_after_audio_session_disappears() {
-        // Corrected behavior (previously `unresolved_browser_does_not_keep_active_meeting_alive`,
-        // which asserted the opposite). A confirmed browser meeting whose URL
-        // evidence lapsed — degrading the live candidate to `UnresolvedBrowser`
-        // with the SAME audio session — must stay Active: the mic is still held,
-        // so the call is still ongoing. The meeting only winds down once the audio
-        // session actually disappears (the user left the call), after which the
-        // normal ending grace fires. This is the keep-alive-on-audio-session fix
-        // for browser calls ending while the user is still in them.
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let active = AudioProcessMeetingState::Active {
-            meeting_id: 123,
-            platform: "Google Meet".to_string(),
-            session_key: key.clone(),
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-            first_seen_at: start,
-            last_seen_at: start,
-            is_browser: true,
-        };
-        let unresolved = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-        let unresolved_candidates = vec![unresolved];
-
-        // URL evidence lapsed but the same browser audio session still holds the
-        // mic → stay Active, no end.
-        let (still_active, action) = advance_audio_process_state(
-            active,
-            &unresolved_candidates,
-            &unresolved_candidates,
-            start + Duration::from_secs(1),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(
-            still_active,
-            AudioProcessMeetingState::Active { .. }
-        ));
-        assert!(action.is_none());
-
-        // Mic released (audio session gone) → transition toward Ending.
-        let (ending, action) = advance_audio_process_state(
-            still_active,
-            &[],
-            &[],
-            start + Duration::from_secs(2),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(ending, AudioProcessMeetingState::Ending { .. }));
-        assert!(action.is_none());
-
-        // Past the ending grace with the session still gone → EndMeeting.
-        let (_idle, action) = advance_audio_process_state(
-            ending,
-            &[],
-            &[],
-            start + Duration::from_secs(23),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(
-            action,
-            Some(AudioProcessStateAction::EndMeeting {
-                meeting_id: 123,
-                ..
-            })
-        ));
-    }
-
-    #[test]
-    fn auto_end_suppresses_same_session_restart_until_audio_session_disappears() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let ending = AudioProcessMeetingState::Ending {
-            meeting_id: 123,
-            platform: "Google Meet".to_string(),
-            session_key: key.clone(),
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-            first_seen_at: start,
-            since: start,
-            is_browser: true,
-        };
-
-        let (idle, action) = advance_audio_process_state(
-            ending,
-            &[],
-            &[],
-            start + Duration::from_secs(21),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(idle, AudioProcessMeetingState::Idle));
-
-        let Some(AudioProcessStateAction::EndMeeting {
-            suppressed_session: Some(session),
-            ..
-        }) = action
-        else {
-            panic!("auto end should carry the session it just ended");
-        };
-
-        let mut suppressed = Vec::new();
-        suppress_session(&mut suppressed, session);
-        let mut candidates = vec![ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/abc-defg-hij".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: key.clone(),
-            first_seen_at: start + Duration::from_secs(22),
-            process: process.clone(),
-        }];
-        filter_suppressed_candidates(&mut candidates, &suppressed);
-        assert!(
-            candidates.is_empty(),
-            "same lingering audio session must not restart the meeting after auto end"
-        );
-
-        let live = live_process_keys(&[process]);
-        retain_live_suppressed_sessions(&mut suppressed, &live);
-        assert!(suppressed.iter().any(|session| session.session_key == key));
-
-        retain_live_suppressed_sessions(&mut suppressed, &HashSet::new());
-        assert!(suppressed.is_empty());
-    }
-
-    #[test]
-    fn explicit_stop_suppresses_current_session_until_process_disappears() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let mut suppressed = vec![SuppressedSession {
-            session_key: key.clone(),
-            platform: Some("Google Meet".to_string()),
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-        }];
-        let mut candidates = vec![ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/abc-defg-hij".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: key.clone(),
-            first_seen_at: start,
-            process: process.clone(),
-        }];
-
-        filter_suppressed_candidates(&mut candidates, &suppressed);
-        assert!(
-            candidates.is_empty(),
-            "same process session should not restart immediately after explicit stop"
-        );
-
-        let live = live_process_keys(&[process]);
-        retain_live_suppressed_sessions(&mut suppressed, &live);
-        assert!(suppressed.iter().any(|session| session.session_key == key));
-
-        retain_live_suppressed_sessions(&mut suppressed, &HashSet::new());
-        assert!(
-            suppressed.is_empty(),
-            "suppression should clear once the audio process session is gone"
-        );
-    }
-
-    #[test]
-    fn explicit_stop_does_not_suppress_new_browser_meeting_url() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        let suppressed = vec![SuppressedSession {
-            session_key: key.clone(),
-            platform: Some("Google Meet".to_string()),
-            meeting_url: Some("https://meet.google.com/old-meet-url".to_string()),
-        }];
-        let mut candidates = vec![ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/new-meet-url".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        }];
-
-        filter_suppressed_candidates(&mut candidates, &suppressed);
-        assert_eq!(
-            candidates.len(),
-            1,
-            "same browser process should be allowed when it resolves to a different meeting URL"
-        );
-    }
-
-    #[test]
-    fn explicit_stop_does_not_suppress_new_audio_session_same_browser_pid() {
-        let old_process = chrome_process();
-        let old_key = ProcessKey::from_process(&old_process).unwrap();
-        let mut new_process = chrome_process();
-        new_process.audio_session_id = Some("coreaudio-process:101:input:built-in-mic".to_string());
-        new_process.audio_object_id = Some(101);
-        let new_key = ProcessKey::from_process(&new_process).unwrap();
-        let start = Instant::now();
-        let suppressed = vec![SuppressedSession {
-            session_key: old_key,
-            platform: Some("Google Meet".to_string()),
-            meeting_url: Some("https://meet.google.com/abc-defg-hij".to_string()),
-        }];
-        let mut candidates = vec![ResolvedMeetingCandidate::Browser {
-            platform: "Google Meet".to_string(),
-            meeting_url: "https://meet.google.com/abc-defg-hij".to_string(),
-            browser_app: "Google Chrome".to_string(),
-            session_key: new_key,
-            first_seen_at: start,
-            process: new_process,
-        }];
-
-        filter_suppressed_candidates(&mut candidates, &suppressed);
-        assert_eq!(
-            candidates.len(),
-            1,
-            "a new audio session should be allowed even when browser pid and URL are reused"
-        );
-    }
-
-    #[tokio::test]
-    async fn active_meeting_blocks_audio_process_insert() {
-        let (_dir, db) = setup_db().await;
-        let active_id = db
-            .insert_meeting("manual", "manual", None, None)
-            .await
-            .unwrap();
-        let manual_meeting = tokio::sync::RwLock::new(None);
-        let outcome =
-            start_or_adopt_auto_meeting(&db, &manual_meeting, "Google Meet", None, None, None)
-                .await;
-        assert_eq!(outcome, AutoStartOutcome::BlockedByActive(active_id));
-
-        let open_count: (i64,) =
-            sqlx::query_as("SELECT COUNT(*) FROM meetings WHERE meeting_end IS NULL")
-                .fetch_one(&db.pool)
-                .await
-                .unwrap();
-        assert_eq!(open_count.0, 1);
-    }
-
-    #[test]
-    fn reattached_meeting_survives_browser_resolution_lag() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let start = Instant::now();
-        // Mimics the post-restart reattach in `run_audio_process_meeting_detection_loop`:
-        // a synthetic `reattached:` key, the platform from the DB row, is_browser:false.
-        let reattached = AudioProcessMeetingState::Active {
-            meeting_id: 42,
-            platform: "Google Meet".to_string(),
-            session_key: ProcessKey::reattached("Google Meet"),
-            meeting_url: None,
-            first_seen_at: start,
-            last_seen_at: start,
-            is_browser: false,
-        };
-        // In the seconds after a restart the browser holding the mic is still
-        // unresolved (DB page evidence / AX walk haven't caught up yet), so it
-        // surfaces as an UnresolvedBrowser (platform "Unknown").
-        let unresolved = ResolvedMeetingCandidate::UnresolvedBrowser {
-            browser_app: "Google Chrome".to_string(),
-            session_key: key,
-            first_seen_at: start,
-            process,
-        };
-        let (state, action) = advance_audio_process_state(
-            reattached,
-            std::slice::from_ref(&unresolved),
-            std::slice::from_ref(&unresolved),
-            start + Duration::from_secs(1),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(
-            matches!(state, AudioProcessMeetingState::Active { .. }),
-            "a reattached meeting must survive the post-restart resolution lag, not drop to Ending"
-        );
-        assert!(action.is_none());
-    }
-
-    #[test]
-    fn reattached_meeting_ends_when_no_session_is_live() {
-        let start = Instant::now();
-        let reattached = AudioProcessMeetingState::Active {
-            meeting_id: 42,
-            platform: "Google Meet".to_string(),
-            session_key: ProcessKey::reattached("Google Meet"),
-            meeting_url: None,
-            first_seen_at: start,
-            last_seen_at: start,
-            is_browser: false,
-        };
-        // No live meeting candidates at all — the reattached meeting must still
-        // wind down (the relaxation only keeps it alive while *some* session is live).
-        let (state, _) = advance_audio_process_state(
-            reattached,
-            &[],
-            &[],
-            start + Duration::from_secs(1),
-            Duration::from_secs(3),
-            Duration::from_secs(20),
-        );
-        assert!(matches!(state, AudioProcessMeetingState::Ending { .. }));
-    }
-
-    #[test]
-    fn ax_resolution_only_runs_before_a_meeting_is_active() {
-        let process = chrome_process();
-        let key = ProcessKey::from_process(&process).unwrap();
-        let now = Instant::now();
-
-        // States where the browser is still being resolved must run the AX walk.
-        assert!(needs_ax_resolution(&AudioProcessMeetingState::Idle));
-        assert!(needs_ax_resolution(&AudioProcessMeetingState::Candidate {
-            platform: "Google Meet".to_string(),
-            session_key: key.clone(),
-            meeting_url: None,
-            first_seen_at: now,
-            is_browser: true,
-        }));
-        assert!(needs_ax_resolution(
-            &AudioProcessMeetingState::CandidateUnresolvedBrowser {
-                browser_app: "Google Chrome".to_string(),
-                session_key: key.clone(),
-                first_seen_at: now,
-                last_resolution_attempt: now,
-            }
-        ));
-
-        // Once a meeting is Active (or winding down), the platform is already
-        // settled — re-walking the AX tree every 1s for the whole call is the
-        // expensive no-op the review flagged.
-        assert!(!needs_ax_resolution(&AudioProcessMeetingState::Active {
-            meeting_id: 1,
-            platform: "Google Meet".to_string(),
-            session_key: key.clone(),
-            meeting_url: None,
-            first_seen_at: now,
-            last_seen_at: now,
-            is_browser: true,
-        }));
-        assert!(!needs_ax_resolution(&AudioProcessMeetingState::Ending {
-            meeting_id: 1,
-            platform: "Google Meet".to_string(),
-            session_key: key,
-            meeting_url: None,
-            first_seen_at: now,
-            since: now,
-            is_browser: true,
-        }));
-    }
-}
+mod tests;
